chore: optimization code
