+++ packages/testing/src/browser/icons/icons.less
@@ -1,23 +1,23 @@
-.sumi_testing_icon_iconFailed {
+.testing_icon_iconFailed {
   color: var(--testing-iconFailed);
 }
 
-.sumi_testing_icon_iconErrored {
+.testing_icon_iconErrored {
   color: var(--testing-iconErrored);
 }
 
-.sumi_testing_icon_iconPassed {
+.testing_icon_iconPassed {
   color: var(--testing-iconPassed);
 }
 
-.sumi_testing_icon_runAction {
+.testing_icon_runAction {
   color: var(--testing-runAction);
 }
 
-.sumi_testing_icon_iconQueued {
+.testing_icon_iconQueued {
   color: var(--testing-iconQueued);
 }
 
-.sumi_testing_icon_iconUnset {
+.testing_icon_iconUnset {
   color: var(--testing-iconUnset);
 }
